Replaced heuristic button ISR with one that instead implements a "leaky integrator" low pass filter, as described at:
https://www.edn.com/a-simple-software-lowpass-filter-suits-embedded-system-applications/

Current values are a sampling rate of 10kHz (every 100 microseconds) and
a k-factor of 2, with a resulting filter bandwidth of 466Hz.  This seems
to provide a good balance between responsiveness and accuracy on my T41.

The button repeat delay logic has also been simplified, and now uses an
elapsedMillis object to provide the same delay without the dependency on
the button ISR, allowing it to be used both with and without button
interrupts enabled.