Added interrupt-driven button thresholds and repeat rate to the EEPROM. Changed software and EEPROM version to T41EEE.1 to reflect newly added fields.
Modified SetAnalogSwitchValues to take advantage of the behavior
provided by the button ISR, when enabled, and theoretically (although
currently untested) also resolve button repeat issues when using the old
code.

Added "Btn Cal" to the calibration menu so that SetAnalogSwitchValues
can be triggered whenever desired.

Added "Btn Repeat" to the calibration menu that allows the interrupt
driven repeat rate to be user-configured.  A value of 0 disables repeat
entirely.