Exposed buttonInterruptsEnabled through SDT.h, marked other global variables related to the button ISR as static within Button.cpp.
Since button repeat is now handled within the ISR, it does not function
for ReadSelectedPushButton in non-interrupt mode, so conditionally call
MyDelay(100L) to restore previous behavior.

Modified SaveAnalogSwitchValues to use original calibration logic for
non-interrupt driven buttons.  This is necessary because of the way the
non-interrupt smoothing works, to ensure that the lowest possible value
is found.  Together with WIGGLE_ROOM this provides the push window that
users would be familiar with from the previous code.

Modified key-up detection in SaveAnalogSwitchValues to work with both
interrupt and non-interrupt driven button code.

Together, these changes allow calibration to be performed properly with
or without button interrupts enabled.  The resulting calibration values
are usable enough that if the mode is changed the previous values work
well enough to use the radio, and certainly well enough to trigger
another button calibration if necessary.